faceid-service/test/requirements_test.txt
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
tqdm==4.66.1
opencv-python-headless>=4.5.0